+++ constants/unicode/unicode.go
@@ -1,7 +1,8 @@
 package unicode
 
 const (
-	ZeroWidthSpace     = "\u200b"
-	InvisibleCharacter = "⁣"
-	NoBreakSpace       = " "
+	ZeroWidthSpace            = "\u200b"
+	InvisibleCharacter        = "⁣"
+	AnotherInvisibleCharacter = "‌" // Zero width non joiner
+	NoBreakSpace              = " "
 )